Move the path making class to a more sensible location
